+++ Tokenizer.h
@@ -12,10 +12,10 @@ namespace Aise {
         
         bool EndOfInput();
         std::shared_ptr<Token> Next();
-        std::shared_ptr<Token> Peek(int offset = 1);
+		std::shared_ptr<Token> Peek(size_t offset = 1);
         
 	private:
         std::vector<std::shared_ptr<Token>> mTokens;
-        int mPosition;
+		size_t mPosition;
 	};
 }
